RectF: clean up some int32->float warnings

+++ src/common/math/RectF.h
@@ -2,6 +2,7 @@
 
 #include "Rect.h"
 
+#include "common/NumericCast.h"
 #include <fmt/core.h>
 #include <stdint.h>
 
@@ -15,7 +16,10 @@ struct RectF {
 	RectF(float x_, float y_, float width_, float height_)
 		: x(x_), y(y_), width(width_), height(height_) {}
 	RectF(Rect other)
-		: x(other.x), y(other.y), width(other.width), height(other.height) {}
+		: x(numeric_cast<float>(other.x)),
+		  y(numeric_cast<float>(other.y)),
+		  width(numeric_cast<float>(other.width)),
+		  height(numeric_cast<float>(other.height)) {}
 
 	float right() const { return x + width; }
 	float bottom() const { return y + height; }
